Set the default name to the TimeStampedName for new datasets
Also better accommodate emtpy aligned bams by recognizing that the mapping
columns may be missing from a pbi file if generated on an emtpy bam. There may
still be some dark corners for emtpy bam support, but for much beyond this and
the pbi file perhaps should be rectified instead.