Add support for merged cmph5 AlignmentSets
This required manipulation of the referenceInfoTable to correct the RowStart
and RowEnd fields, which required reordering the index for cmph5 based
AlignmentSets to make the AlignmentSet as a whole conform to the ordering of a
cmph5. As an added bonus it is now trivial to do the same thing to a bam,
should we choose to in the future. There were some other minor issues uncovered
during the process that have also been resolved, mostly more aggressive cache
clearing and better filter support. Also updated some documentation.