Speed up cmp.h5 reader initialization in the case of many references
- Avoid indexing H5 datasets directly in inner loops---precopy to
  memory
- Avoid slow recarray indexing in inner loops
- Don't repeatedly access the offsettable from the file

This brings init time down from >12 seconds to 4 seconds on a cmp.h5
file with 4096 reference contigs.  It's still slow--the remaining
slowness is attributable to the aggressive pre-caching of the
reference "groups", which could be replaced by a lazier scheme if we
want.

This was a use case that I personally didn't encounter often in the
past, so it never was optimized...

[Tracking bug 24842]