Add tests, add referenceInfo(ID), clean up code
Add tests for multiple resources, some with the same reference, one with a
different reference. No duplicates are expected.

Add the option to pass a reference ID to referenceInfo. This is inferior to
using the refernece Name as a unique identifier, but good for our older GC
users.

Due to the recent changes to the referenceInfoTable and index table, a lot of
code can be spruced up. This is a start