Raise an exception on missing ExternalResource files
This is a new default behavior, can be overridden using skipMissing=True.
Had to add this to a bunch of tests that use the example datasets.

Also cleaned up the way that assertIndexed works. No more of this "strict"
dance. It was no longer necessary and often resulted in files being opened
multiple times unnecessarily.