Expand performance improvements to alternative writeRecord syntax
It turns out that the expensive part of creating a FastaRecord using an
MmappedFastaSequence was the set of "assert not in" statements. The
__contains__ provided by Sequence wasn't particularly efficient for our use
case. Instead of accepting the performance hit, passing in the sequence as a
string when casting the record, or implementing a more efficient __contains__,
we can leave the record argument as an IndexedFastaRecord (now with a __str__),
which is fast enough, or create a FastaRecord from a (header, str(sequence)),
which should be fast and memory efficient enough for contigs < 2GB in size.