Add new "future" partnumbers to pbcore See: https://bugzilla.nanofluidics.com/show_bug.cgi?id=31748
The goal here is to support new chemistry releases out of sync with
software, so long as new trainings/parameters are not required.

To make this work, systems engineering and project management played
out several scenarios.  The scenario here is that we will have new
BK/SK releases which do not require new trainings---2.x/2.y will
resolve as "S/P2-C2/prospective" and CCS, etc. will use the S/P1-C1
training for that condition.

If we determine that a new training is required, *different*
partnumbers will be used.  The user experienece will then be:
secondary will fail to resolve the chemistry, and the user will be
informed to check for a SW update.

In a future SW release, once the actual S/P2-C2 has played out, we
will revisit these mappings.  If in fact the actual S/P2-C2 chemistry
requires a new training, we will have issued different partnumbers;
those will then be inserted here.  One way or the other, the
"prospective" designation will be removed here and in client code.

One other thing that is happening here: the pre-Echidna-era chemistry
is now being relabeled as S/P1-C1/beta, *even* for basecaller 3.1+.