[DEP-5443] use sl-pypi-local
The necessary artifacts here were populated by `pip download`ing the
output of `pip freeze` into a temporary directory, then uploading that
set of wheels straight into `sl-pypi-local/`.

Since this local repository has a variety of package versions for
different purposes, we'll also explicitly lock the versions of
packages we use with a constraint.lock file. This ensures that, while
we've unpinned dependencies in other parts of the build system, we'll
still be consitently using the same set of dependencies.

Signed-off-by: Jake Helmert <jhelmert@pacificbiosciences.com>